refactor: PathValidator 로직 Piece로 이동
- 단순히 PathValidator는 위임만 하고 있으므로, 데이터를 가지고 있는 Piece에서 수행하는 것이 더 명확하고 구조가 간단하다.
